End of section 5

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -9,6 +9,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -52,12 +53,13 @@ namespace AuctionService.Controllers
             }
             return _mapper.Map<AuctionDto>(auction);
         }
-
+        
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
         {
             var auction = _mapper.Map<Auction>(createAuctionDto);
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
             _context.Auctions.Add(auction);
             var newAuction = _mapper.Map<AuctionDto>(auction);
             await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
@@ -68,6 +70,7 @@ namespace AuctionService.Controllers
             }
             return CreatedAtAction(nameof(GetAuctionById), new {auction.Id }, _mapper.Map<AuctionDto>(auction));
         }
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -76,6 +79,11 @@ namespace AuctionService.Controllers
             if (auction == null)
             {
                 return NotFound();
+            }
+
+            if (auction.Seller != User.Identity.Name)
+            {
+                return Forbid();
             }
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -90,13 +98,18 @@ namespace AuctionService.Controllers
             return BadRequest("Failed to update auction");
         }
 
-         [HttpDelete("{id}")]
+        [Authorize]
+        [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
             if (auction == null)
             {
                 return NotFound();
+            }
+            if (auction.Seller != User.Identity.Name)
+            {
+                return Forbid();
             }
             _context.Auctions.Remove(auction);
             await _publishEndpoint.Publish(new AuctionDeleted { Id = auction.Id.ToString() });
